Resolved cast before assignment.
Sonarqube:
http://nemo.sonarqube.org/issues/search#issues=20152c49-b33a-462e-b4d7-351ea150a1f3

+++ mr/src/main/java/org/apache/mahout/cf/taste/impl/recommender/svd/ParallelSGDFactorizer.java
@@ -227,7 +227,7 @@ public class ParallelSGDFactorizer extends AbstractFactorizer {
 
   //TODO: needs optimization
   private double getMu(int i) {
-    return mu0 * Math.pow(decayFactor, i - 1) * Math.pow(i + stepOffset, forgettingExponent);
+    return (double)mu0 * Math.pow(decayFactor, i - 1) * Math.pow(i + stepOffset, forgettingExponent);
   }
 
   @Override
